requirements.txt
# Core dependencies
python-dotenv>=0.19.0
openai>=1.0.0

# Flask and extensions
flask>=2.0.0
flask-cors>=3.0.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0

# Utilities
requests>=2.31.0